Update authService.js
-Implemented JWT
-Added validation and error handling
-Added account method

+++ server/src/services/authService.js
@@ -1,5 +1,7 @@
 const UserRepository = require('../repositories/userRepository');
+const { BadRequestError, UnauthorizedError } = require('../utils/customErrors');
 const bcrypt = require('bcrypt');
+const { generateToken } = require('../utils/authHelpers');
 // const jwt = require('jsonwebtoken');
 
 const authService = {
@@ -11,10 +13,11 @@ const authService = {
      * @returns {Promise<User>}
      */
     async register(username, email, password) {
+        if (!username || !email || !password) throw new BadRequestError('Missing fields');
+
         const existingUser = await UserRepository.getByEmail(email);
-        if (existingUser) {
-            throw new Error('User already exists');
-        }
+        if (existingUser) throw new BadRequestError('User already exists');
+
         const hashedPassword = await bcrypt.hash(password, 10);
 
         return await UserRepository.create(username, email, hashedPassword);
@@ -27,18 +30,14 @@ const authService = {
      */
     async login(email, password) {
         const user = await UserRepository.getByEmail(email);
-        if (!user) {
-            throw new Error('Invalid email or password');
-        }
+        if (!user) throw new UnauthorizedError('Invalid email or password');
 
-        const isPasswordValid = await user.veirfyPassword(password, bcrypt);
-        if (!isPasswordValid) {
-            throw new Error('Invalid email or password');
-        }
+        const isPasswordValid = await user.verifyPassword(password, bcrypt);
+        if (!isPasswordValid) throw new UnauthorizedError('Invalid email or password');
 
-        //JWT to-do
+        const token = generateToken(user);
 
-        return user.toPublic();
+        return { user, token };
     },
 
     /**
@@ -47,12 +46,25 @@ const authService = {
      * @returns {Promise<User>}
      */
     async update(userId, updates) {
+        if (updates.length === 0) throw new BadRequestError('No fields to update');
+
         if (updates.password) {
             updates.passwordHash = await bcrypt.hash(updates.password, 10);
             delete updates.password;
         }
 
-        return await UserRepository.update(userId, updates).toPublic();
+        return await UserRepository.update(userId, updates);
+    },
+
+    /**
+     * 
+     * @param {int} userId 
+     * @returns 
+     */
+    async account(userId) {
+        if (!userId) throw new BadRequestError('User ID is required');
+
+        return await UserRepository.getById(userId);
     },
 
     /**
@@ -60,6 +72,8 @@ const authService = {
      * @returns {Promise<boolean>}
      */
     async delete(userId) {
+        if (!userId) throw new BadRequestError('User ID is required');
+
         return await UserRepository.delete(userId);
     }
